Renamed update-viewState to trips to ensure consistency
