+++ examples/trips/package.json
