+++ examples/trips/scene.js
