+++ examples/trips/webpack.config.js
